+++ surepy/analysis/accumulation_analysis.py
@@ -218,18 +218,18 @@ class AccumulationClusterCheck(_Analysis):
 
     def plot(self, ax=None, **kwargs):
         """
-        Provide plot of results as matplotlib.axes.Axes object.
+        Provide plot of results as :class:`matplotlib.axes.Axes` object.
 
         Parameters
         ----------
-        ax : matplotlib.axes.Axes
+        ax : :class:`matplotlib.axes.Axes`
             The axes on which to show the image
         kwargs : dict
-            Other parameters passed to `matplotlib.pyplot.plot()`.
+            Other parameters passed to :func:`matplotlib.pyplot.plot`.
 
         Returns
         -------
-        matplotlib.axes.Axes
+        :class:`matplotlib.axes.Axes`
             Axes object with the plot.
         """
         if ax is None: