+++ surepy/tests/test_locdata_io.py
@@ -17,7 +17,7 @@ def test_convert_property_types(locdata_2d):
     df = locdata_2d.data.copy()
     types_mapping = {'position_x': 'float', 'position_y': str, 'frame': np.int64, 'not_in_there': 'float'}
     converted_df = io.convert_property_types(dataframe=df, loc_properties=None, types=types_mapping)
-    assert all(converted_df.dtypes == [np.float32, object, np.int64, np.int32])
+    assert all(converted_df.dtypes == [np.float32, object, np.int64, int])
     assert isinstance(converted_df['position_y'].iloc[0], str)